requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
python-dotenv>=1.0
tensorflow
google-generativeai
celery>=5.3
redis>=5.0
pytest